Proposed changes to API
