+++ .gitignore
@@ -12,7 +12,7 @@ tsconfig.tsbuildinfo
 /colors
 /fonts
 /icons
-/Form
+/TextField
 /Modal
 /Card
 /shared
