--- src/Form/index.ts
@@ -1,2 +1,0 @@
-export * from "./FormField";
-export * from "./Stepper";
